feat(): Added a key_preview layout.
The added key_preview layout is referenced by keyboard_view.xml in its `android:keyPreviewLayout` attribute.
The key preview shows up when one presses their finger on a key (i.e. to better indicate the currently pressed key).
I believe it is a part of the official (deprecated) implementation of `KeyboardView` and `Keyboard`.

Refs:
- https://www.androidauthority.com/lets-build-custom-keyboard-android-832362/
- https://www.youtube.com/watch?v=UHc3PN4l850