feat(.WhisperInputService): Added interface inheritance (KeyboardView.OnKeyboardActionListener).
However, the KeyboardView class is deprecated in API Level 29. We shall focus on making it work before diving into resolving this issue.
In future branches, it is planned to either copy this implementation into this project (as [the official suggestion](https://developer.android.com/reference/android/inputmethodservice/KeyboardView)) or to implement a new, customized keyboard system.

Refs:
- https://developer.android.com/reference/android/inputmethodservice/KeyboardView
- https://www.androidauthority.com/lets-build-custom-keyboard-android-832362/
- https://www.youtube.com/watch?v=UHc3PN4l850