feat(): Ripples now reflect real-time microphone amplitudes.
Images were updated. Original version did not meet expectations. As with previous commits, they are made with draw.io by @tigerpaws01.
Microphone amplitude is tested on emulator to be in the range from 0 to 32767.
Microphone amplitudes are periodically reported by RecorderManager, captured by WhisperInputService and passed to WHisperKeyboard for animation display.

Refs:
- https://stackoverflow.com/questions/37281246/animate-imageview-from-alpha-0-to-1
- https://stackoverflow.com/questions/13629102/android-how-to-convert-getmaxamplitude-to-db